requirements.txt
requests
pyyaml
pixels2svg
Pillow
